Adding Inference.hs providing typecheck :: Program -> Stage Program to perform type-inference
Largely based of the implementation of Hindley-Milner described in
"The Implementation of Functional Programming Languages" with some help
from "Typing Haskell in Haskell".

Currently, we convert the source program to a single letrec. For example,

data List a = Cons a (List a) | Nil;

length ls = case ls of {
    Cons x xs -> 1 + length ls;
    Nil       -> Nil;
};

main = length (Cons 1 (Cons 2 (Cons 3 Nil)));

gets transformed to

data List a = Cons a (List a) | Nil;
letrec {
    length = \ls -> case ls of {
        Cons x xs -> 1 + length ls;
        Nil       -> Nil;
    };
    main = length (Cons 1 (Cons 2 (Cons 3 Nil)));
} in main;

and then typechecked as a single expression. I might revisit this since it
makes giving good error messages difficult.