Adding Transforms/StronglyConnectedComponents.hs
In "Implementing Functional Languages" this is used to implement full
laziness. I'm using it to transform a program like this:

fix f = f (fix f);
id x = x;
main = fix id;

into a single let(rec) like this:

letrec {
    fix = \f -> f (fix f);
} in let {
    id = \x -> x;
} in let {
    main = fix id;
} in main;

for type-checking. This ensures that we maximize polymorphism. Initially, we
would just put everything into a single letrec like so:

letrec {
    fix = \f -> f (fix f);
    id = \x -> x;
    main = fix id;
} in main;

but then id is monomorphic in the definition of main.